Adjusted G711Codec to disambiguate Math.Min by casting to int before comparison, resolving ambiguous overload between short and int.

X-Lovable-Edit-ID: edt-8dc7a647-8f73-4e70-8cd2-be7e4508350b

+++ csharp-sip-bridge/AdaVaxVoIP/Audio/G711Codec.cs
@@ -59,7 +59,7 @@ public static class G711Codec
             mask = 0xFF;
         }
 
-        pcm = (short)Math.Min(pcm, cClip);
+        pcm = (short)Math.Min((int)pcm, (int)cClip);
         pcm += 0x84;
 
         int seg;
